+++ e2e/pages/testrun/list.po/constants.ts
@@ -12,6 +12,7 @@ export const names = {
 };
 
 export const columns = {
-    build: 'Build'
+    build: 'Build',
+    milestone: 'Milestone'
 };