+++ src/app/shared/models/error.ts
@@ -1,0 +1,3 @@
+export class ATError {
+    message: string;
+}