+++ src/app/shared/models/import.ts
@@ -2,7 +2,7 @@ export class Import {
   id?: number;
   testrun_id?: number;
   project_id?: number;
-  is_finished?: number;
+  finish_status?: number;
   started?: Date|String|number;
   finished?: Date|String|number;
   log?: string;